third_party/WebKit/LayoutTests/external/wpt/page-visibility/idlharness-expected.txt

This is a testharness.js-based test.
PASS Document interface: attribute hidden 
PASS Document interface: attribute visibilityState 
FAIL Document interface: attribute onvisibilitychange assert_true: The prototype object must have a property "onvisibilitychange" expected true got false
FAIL Document must be primary interface of window.document assert_equals: window.document's prototype is not Document.prototype expected object "[object Document]" but got object "[object HTMLDocument]"
FAIL Stringification of window.document assert_equals: class string of window.document expected "[object Document]" but got "[object HTMLDocument]"
PASS Document interface: window.document must inherit property "hidden" with the proper type (0) 
PASS Document interface: window.document must inherit property "visibilityState" with the proper type (1) 
FAIL Document interface: window.document must inherit property "onvisibilitychange" with the proper type (2) assert_inherits: property "onvisibilitychange" not found in prototype chain
Harness: the test ran to completion.
